+++ examples/github_thread.rb
@@ -2,12 +2,10 @@
 
 $LOAD_PATH << File.expand_path('../lib', __dir__)
 
+require 'logger'
 require 'restify'
 
-if ENV['LOGGING']
-  Logging.logger.root.add_appenders Logging.appenders.stdout
-  Logging.logger.root.level = :debug
-end
+Restify.logger = Logger.new($stdout, level: :debug) if ENV['LOGGING']
 
 if (token = ENV.fetch('GITHUB_TOKEN', nil))
   headers['Authorization'] = "Bearer #{token}"